requirements.txt
albumentations
opencv-contrib-python
jupyter
matplotlib
lightning
wandb
einops
pandas
scikit-learn
datasets
transformers
kornia
bitsandbytes
peft
timm
openvino
nncf
lightgbm
triton
torchmetrics
ultralytics
heavyball